feat(pci.db): add capabilities on user edit and rename files
