+++ packages/manager/apps/pci-databases-analytics/src/pages/services/[serviceId]/users/_components/deleteUser.tsx
@@ -41,7 +41,7 @@ const DeleteUser = ({
       toast.toast({
         title: t('deleteUserToastErrorTitle'),
         variant: 'destructive',
-        description: err.message,
+        description: err.response.data.message,
       });
       if (onError) {
         onError(err);
